+++ Database/models/ExascaleDbStorageVault.cs
@@ -153,6 +153,12 @@ namespace Oci.DatabaseService.Models
         public System.Collections.Generic.Dictionary<string, System.Collections.Generic.Dictionary<string, System.Object>> DefinedTags { get; set; }
         
         /// <value>
+        /// The [OCID](https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) of the Exadata infrastructure.
+        /// </value>
+        [JsonProperty(PropertyName = "exadataInfrastructureId")]
+        public string ExadataInfrastructureId { get; set; }
+        
+        /// <value>
         /// System tags for this resource. Each key is predefined and scoped to a namespace.
         /// For more information, see [Resource Tags](https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
         ///